Move config one level up from `lib`
